refactor/#23
Building and Running now works. We'll keep the default naming scheme setup: `my-app/src/X/java/com/mycompany/app/` to avoid complications.
