+++ my-app/pom.xml
@@ -4,11 +4,11 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>DD2480.group16.decide</groupId>
-  <artifactId>decide</artifactId>
+  <groupId>com.mycompany.app</groupId>
+  <artifactId>my-app</artifactId>
   <version>1.0-SNAPSHOT</version>
 
-  <name>decide</name>
+  <name>my-app</name>
   <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
