+++ my-app/src/test/java/com/mycompany/app/AppTest.java
@@ -1,4 +1,4 @@
-package DD2480.group16.decide;
+package com.mycompany.app;
 
 import static org.junit.Assert.assertTrue;
 
@@ -7,7 +7,7 @@ import org.junit.Test;
 /**
  * Unit test for simple App.
  */
-public class ProjectTest 
+public class AppTest 
 {
     /**
      * Rigorous Test :-)
